queue/worker/requirements.txt
aiosqlite==0.20.0
gitpython==3.1.43
neo4j==5.23.1
prometheus-client==0.20.0
python-json-logger==2.0.7
